+++ client/web/src/storm/pages/SearchPage/SearchPageContent.tsx
@@ -35,7 +35,6 @@ export const SearchPageContent: FC<SearchPageContentProps> = props => {
     const isLightTheme = useIsLightTheme()
     const [experimentalQueryInput] = useExperimentalQueryInput()
     const [ownFeatureFlagEnabled] = useFeatureFlag('search-ownership')
-    const [codySearchEnabled] = useFeatureFlag('cody-experimental')
     const enableOwnershipSearch = ownEnabled && ownFeatureFlagEnabled
 
     /** The value entered by the user in the query input */
@@ -87,7 +86,7 @@ export const SearchPageContent: FC<SearchPageContentProps> = props => {
                 ) : (
                     <>
                         <SearchPageInput queryState={queryState} setQueryState={setQueryState} />
-                        {!codySearchEnabled && !authenticatedUser && isSourcegraphDotCom && (
+                        {!window.context.codyEnabled && !authenticatedUser && isSourcegraphDotCom && (
                             <div className="d-flex justify-content-center mt-4">
                                 <Text className="text-muted">
                                     <Badge variant="merged">Experimental</Badge>{' '}
@@ -105,7 +104,7 @@ export const SearchPageContent: FC<SearchPageContentProps> = props => {
                                 </Text>
                             </div>
                         )}
-                        {codySearchEnabled && (
+                        {window.context.codyEnabled && (
                             <div className="d-flex justify-content-center mt-4">
                                 <Text className="text-muted">
                                     <Badge variant="merged">Experimental</Badge>{' '}